+++ src/tests/api/test_dataset.py
@@ -6,7 +6,7 @@ from api.viewsets import Dataset, Project
 
 
 @pytest.mark.django_db
-class TestUserEndpoints:
+class TestDatasetEndpoints:
     endpoint = '/api/v1/dataset'
     unique_fields = 'The fields user, name must make a unique set.'
 
@@ -24,24 +24,19 @@ class TestUserEndpoints:
         response = client.get(f"{self.endpoint}?project={project.pk}")
         assert response.status_code == 403
 
-    def test_create_project(self, client_as_user: APIClient):
-        body = {"name": self.PROJECT_NAME, "user": str(client_as_user[1].uuid)}
-        response = client_as_user[0].post(f"{self.endpoint}", body, format='json')
-        assert response.status_code == 201
-
-    def test_create_project(self, client_as_user: APIClient, project: Project):
+    def test_create_dataset(self, client_as_user: APIClient, project: Project):
         body = {"name": self.DATASET_NAME, "user": str(client_as_user[1].uuid), "project": project.pk}
         response = client_as_user[0].post(f"{self.endpoint}", body, format='json')
         assert response.status_code == 201
 
-    def test_create_project_with_existing_name(self, client_as_user: APIClient, project: Project):
+    def test_create_dataset_with_existing_name(self, client_as_user: APIClient, project: Project):
         baker.make(Dataset, name=self.DATASET_NAME, user=client_as_user[1])
         body = {"name": self.DATASET_NAME, "user": str(client_as_user[1].uuid), "project": project.pk}
         response = client_as_user[0].post(f"{self.endpoint}", body, format='json')
         assert response.status_code == 400
         assert response.json()['non_field_errors'][0] == self.unique_fields
 
-    def test_create_project_without_permissions(self, client: APIClient, project: Project):
+    def test_create_dataset_without_permissions(self, client: APIClient, project: Project):
         body = {"name": self.DATASET_NAME, "user": 'abcd-efg', "project": project.pk}
         response = client.post(f"{self.endpoint}", body, format='json')
         assert response.status_code == 403
@@ -50,6 +45,11 @@ class TestUserEndpoints:
         response = client_as_admin[0].put(f"{self.endpoint}")
         assert response.status_code == 405
 
-    def test_delete(self, client_as_admin: APIClient):
-        response = client_as_admin[0].delete(f"{self.endpoint}")
-        assert response.status_code == 405
+    def test_delete(self, client_as_admin: APIClient, dataset: Dataset):
+        response = client_as_admin[0].delete(f"{self.endpoint}/{dataset.pk}")
+        assert response.status_code == 204
+
+    def test_delete_not_existing(self, client_as_admin: APIClient):
+        dataset = baker.prepare(Dataset)
+        response = client_as_admin[0].delete(f"{self.endpoint}/{dataset.pk}")
+        assert response.status_code == 404