+++ src/tests/api/test_project.py
@@ -6,7 +6,7 @@ from api.viewsets import Project
 
 
 @pytest.mark.django_db
-class TestUserEndpoints:
+class TestProjectEndpoints:
     endpoint = '/api/v1/project'
     unique_fields = 'The fields user, name must make a unique set.'
 
@@ -45,6 +45,11 @@ class TestUserEndpoints:
         response = client_as_admin[0].put(f"{self.endpoint}")
         assert response.status_code == 405
 
-    def test_delete(self, client_as_admin: APIClient):
-        response = client_as_admin[0].delete(f"{self.endpoint}")
-        assert response.status_code == 405
+    def test_delete(self, client_as_admin: APIClient, project: Project):
+        response = client_as_admin[0].delete(f"{self.endpoint}/{project.pk}")
+        assert response.status_code == 204
+
+    def test_delete_not_existing(self, client_as_admin: APIClient):
+        project = baker.prepare(Project)
+        response = client_as_admin[0].delete(f"{self.endpoint}/{project.pk}")
+        assert response.status_code == 404